+++ app/features/expenses/components/table.tsx
@@ -31,6 +31,7 @@ export default function ExpenseTable() {
 
 	const onDelete = useCallback(
 		async (id: string) => {
+			if (!confirm('Apakah Anda yakin ingin menghapus data ini?')) return;
 			try {
 				await deleteExpense(id);
 				toast.success(messages.deleted);
